waiting modify image

+++ back-end/controllers/sauce.js
@@ -4,7 +4,6 @@ const fs = require("fs");
 exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
   delete sauceObject._id;
-  // delete sauceObject._userId;
   const sauce = new Sauce({
     ...sauceObject,
     userId: req.auth.userId,
@@ -43,42 +42,42 @@ exports.getOneSauce = (req, res, next) => {
 
 exports.likesDislikesSauces = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id }).then((sauce) => {
-    if (!sauce.usersLiked.includes(req.body.userId) && req.body.like === 1) {
+    if (!sauce.usersLiked.includes(req.auth.userId) && req.body.like === 1) {
       Sauce.updateOne(
         { _id: req.params.id },
-        { $inc: { likes: 1 }, $push: { usersLiked: req.body.userId } }
+        { $inc: { likes: 1 }, $push: { usersLiked: req.auth.userId } }
       )
         .then(() => res.status(200).json({ message: "sauce liked !" }))
         .catch((error) => res.status(400).json({ error }));
     } else if (
-      !sauce.usersDisliked.includes(req.body.userId) &&
+      !sauce.usersDisliked.includes(req.auth.userId) &&
       req.body.like === -1
     ) {
       Sauce.updateOne(
         { _id: req.params.id },
-        { $inc: { dislikes: 1 }, $push: { usersDisliked: req.body.userId } }
+        { $inc: { dislikes: 1 }, $push: { usersDisliked: req.auth.userId } }
       )
         .then(() => res.status(200).json({ message: "sauce disliked !" }))
         .catch((error) => res.status(400).json({ error }));
     } else if (
-      sauce.usersLiked.includes(req.body.userId) &&
+      sauce.usersLiked.includes(req.auth.userId) &&
       req.body.like === 0
     ) {
       Sauce.updateOne(
         { _id: req.params.id },
-        { $inc: { likes: -1 }, $pull: { usersLiked: req.body.userId } }
+        { $inc: { likes: -1 }, $pull: { usersLiked: req.auth.userId } }
       )
         .then(() =>
           res.status(200).json({ message: "Like removed from sauce !" })
         )
         .catch((error) => res.status(400).json({ error }));
     } else if (
-      sauce.usersDisliked.includes(req.body.userId) &&
+      sauce.usersDisliked.includes(req.auth.userId) &&
       req.body.like === 0
     ) {
       Sauce.updateOne(
         { _id: req.params.id },
-        { $inc: { dislikes: -1 }, $pull: { usersDisliked: req.body.userId } }
+        { $inc: { dislikes: -1 }, $pull: { usersDisliked: req.auth.userId } }
       )
         .then(() =>
           res.status(200).json({ message: "Dislike removed from sauce !" })
@@ -105,24 +104,25 @@ exports.modifySauce = (req, res, next) => {
         description: req.body.description,
         mainPepper: req.body.mainPepper,
         heat: req.body.heat,
-      }; //modifier req.body
+      };
 
   delete sauceObject._userId;
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
       if (sauce.userId != req.auth.userId) {
-        //middleware pour 106 avant multer
         res.status(401).json({ message: "Not authorized" });
       } else {
-        const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          Sauce.updateOne(
-            { _id: req.params.id },
-            { ...sauceObject, _id: req.params.id }
-          )
-            .then(() => res.status(200).json({ message: "Objet modifié!" }))
-            .catch((error) => res.status(401).json({ error }));
-        });
+        if (req.file && sauce.imageUrl) {
+          const filename = sauce.imageUrl.split("/images/")[1];
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.updateOne(
+              { _id: req.params.id },
+              { ...sauceObject, _id: req.params.id }
+            )
+              .then(() => res.status(200).json({ message: "Objet modifié!" }))
+              .catch((error) => res.status(401).json({ error }));
+          });
+        }
       }
     })
     .catch((error) => {
